ft-components/src/main/java/uk/co/slc/sfd/cp/ft/jsf/converter/TrimWhiteSpaceConverter.java
package uk.co.slc.sfd.cp.ft.jsf.converter;

import javax.faces.component.UIComponent;
import javax.faces.context.FacesContext;
import javax.faces.convert.Converter;
import javax.faces.convert.FacesConverter;

/**
@author slc
 * 
 * JSF converter to remove white space from an input
 * 
 */
@FacesConverter(value = "trimWhiteSpaceConverter")
public class TrimWhiteSpaceConverter implements Converter {

    /*
     * (non-Javadoc)
     * 
     * @see
     * javax.faces.convert.Converter#getAsObject(javax.faces.context.FacesContext
     * , javax.faces.component.UIComponent, java.lang.String)
     */
    @Override
    public Object getAsObject(FacesContext context, UIComponent component, String value) {

        return value == null ? null : value.trim();
    }

    /*
     * (non-Javadoc)
     * 
     * @see
     * javax.faces.convert.Converter#getAsString(javax.faces.context.FacesContext
     * , javax.faces.component.UIComponent, java.lang.Object)
     */
    @Override
    public String getAsString(FacesContext context, UIComponent component, Object object) {

        return (String) object;
    }

}
